feat: Add Exercise Dropdown and Editable Exercise Table Components
- Implemented ExerciseDropdown component for searchable exercise selection.
- Created ExerciseEditRow component for inline editing of exercise properties.
- Added ExerciseTableHeader for defining the structure of the exercise table.
- Refactored ExerciseTableInline component to support inline editing with new components.
- Introduced ExerciseTableRow for displaying exercise data with edit and delete actions.
- Added utility functions for deriving unique motions and target areas, and calculating TUT.
- Updated phase processors to improve exercise ID handling.
- Cleaned up imports in workout queue integration for better code organization.
